% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caemovir-package.R
\docType{package}
\name{caemovir-package}
\alias{caemovir}
\alias{caemovir-package}
\title{caemovir: Agent-based Epidemiological Model Of Victoria In C and R}
\description{
A model of Victoria Australia, for the purposes of measuring spread of a disease.
}
\author{
\strong{Maintainer}: Hugh Parsonage \email{hugh.parsonage@gmail.com}

}
\keyword{internal}
