% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_seq.R
\name{check_enseq}
\alias{check_enseq}
\title{Check whether a vector is fully within a sequence}
\usage{
check_enseq(x, m = max(x, na.rm = TRUE))
}
\arguments{
\item{x}{An integer vector.}

\item{m}{An integer, the maximum value of \code{x}.}
}
\value{
A vector is \strong{ensequenced} if it contains every integer between 1 and
its maximum. It is \code{m}-\strong{ensequenced} if its maximum is \code{m}.



\code{check_seq} returns \code{0} if \code{x} is ensequenced; otherwise the
value which violates the conditions (either an element of \code{x} not between
\code{1} and \code{m} or the absent value.)
}
\description{
When passing a representation of some identifiers to a C function,
 it is convenient to be able to assume that the identifiers are elements of
 the sequence \code{1:N} for some \code{N}.
}
\examples{
x <- sample.int(100)
check_enseq(x)
check_enseq(x, m = 99L)
check_enseq(x, m = 101L)
x <- sample.int(100, 99)
check_enseq(x)  # missing value

}
