% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse}
\alias{collapse}
\title{Collapse a vector into its nonzero elements}
\usage{
collapse(x)
}
\arguments{
\item{x}{An integer vector.}
}
\value{
\code{x[x > 0]}
}
\description{
Collapse a vector into its nonzero elements
}
