% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample_fixed_TRUE}
\alias{sample_fixed_TRUE}
\title{Sample to give a fixed number of TRUE values}
\usage{
sample_fixed_TRUE(n, nTRUE)
}
\arguments{
\item{n}{An integer, the length of the output}

\item{nTRUE}{The number of \code{TRUE} values in the result.}
}
\value{
A logical vector of length \code{n} with \code{nTRUE} \code{TRUE}
elements interspersed randomly throughout.
}
\description{
Sample to give a fixed number of TRUE values
}
