% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_vic.R
\name{simulate_vic}
\alias{simulate_vic}
\title{Simulate VIC}
\usage{
simulate_vic(
  days_to_simulate = 28,
  PolicyPars = set_policypars(),
  EpiPars = set_epipars(),
  InitialStatus = NULL,
  FutureStatus = NULL,
  showProgress = getOption("caemovir.showProgress", 1L),
  nThread = getOption("caemovir.nThread", 1L)
)
}
\arguments{
\item{days_to_simulate}{}
}
\description{
Simulate VIC
}
